// frontend/lib/generated-api/models/DocumentIssueOutput.ts
export interface DocumentIssueOutput {
  /**
   * The title of the issue
   * @type {string}
   * @memberof DocumentIssueOutput
   */
  title: string;
  /**
   * The description of the issue
   * @type {string}
   * @memberof DocumentIssueOutput
   */
  description: string;
  /**
   * The severity of the issue
   * @type {SeverityEnum}
   * @memberof DocumentIssueOutput
   */
  severity: SeverityEnum;
  /**
   *
   * @type {string}
   * @memberof DocumentIssueOutput
   */
  additionalContext?: string | null;
  /**
   * Custom context data with structured label-value pairs, including more details about the issue and/or context of the issue
   * @type {Array<ContextDataOutput>}
   * @memberof DocumentIssueOutput
   */
  contextData?: Array<ContextDataOutput>;
  /**
   *
   * @type {number}
   * @memberof DocumentIssueOutput
   */
  chunkIndex?: number | null;
  /**
   *
   * @type {number}
   * @memberof DocumentIssueOutput
   */
  claimIndex?: number | null;
  /**
   *
   * @type {ClaimCategory}
   * @memberof DocumentIssueOutput
   */
  claimCategory?: ClaimCategory | null;
}